Bugfixing in LogginPolicy
git-svn-id: https://ecell.svn.sourceforge.net/svnroot/ecell/ecell3/trunk@2136 f1531174-cb10-0410-9fe6-89aa7ac3eedb

+++ ecell/libemc/LocalSimulatorImplementation.cpp
@@ -309,7 +309,7 @@ namespace libemc
   createLogger( libecs::StringCref aFullPNString )
   {
    libecs::PolymorphVector aVector;
-   aVector.push_back(Integer(0));
+   aVector.push_back(Integer(1));
    aVector.push_back(Real(0.0));
    aVector.push_back(Integer(0));
    aVector.push_back(Integer(0));
